解决 legacy 问题
