+++ components/alert/alert-legacy/__test__/index.test.js
