+++ components/alert/alert-legacy/style/index.js
