+++ components/alert/alert-legacy/style/index.scss
@@ -1,6 +1,6 @@
 @import '../../style/index.scss';
 
-.hi-alert {
+.hi-alert-legacy {
   position: relative;
   padding: 12px;
   background: #ebf2fd;
