+++ components/badge/badge-legacy/style/index.scss
@@ -1,4 +1,4 @@
-.hi-badge {
+.hi-badge-legacy {
   &-base {
     display: inline-block;
     position: relative;
