+++ components/index.js
@@ -6,9 +6,7 @@ export { default as Table } from './table'
 export { default as Notification } from './notification'
 export { handleNotificate } from './notification'
 export { default as Modal } from './modal'
-// export { handleModal } from './modal'
 export { default as Alert } from './alert'
-// export { handleAlert } from './alert'
 export { default as Panel } from './panel'
 export { default as Collapse } from './collapse'
 export { default as Select } from './select'
