+++ components/progress/progress-legacy/CircleProgress.js
@@ -1,7 +1,7 @@
 import React from 'react'
 export const CircleProgress = (props) => {
   let prefix = 'hi-progress'
-  const {percent: percentNum, text, status, radius: radiusThis, withOutText = false} = props
+  const { percent: percentNum, text, status, radius: radiusThis, withOutText = false } = props
   const percent = percentNum > 0 ? percentNum : 0
   const radius = radiusThis > 0 ? radiusThis : 40
   const totalRadiusWidth = radius + 10
@@ -9,7 +9,7 @@ export const CircleProgress = (props) => {
   return (
     <div
       className={`${prefix}__svg`}
-      style={{width: radius * 2, height: radius * 2}}>
+      style={{ width: radius * 2, height: radius * 2 }}>
       <svg
         viewBox={`0 0 ${totalRadiusWidth * 2} ${totalRadiusWidth * 2}`}>
         <circle
