+++ components/progress/progress-legacy/DashboardProgress.js
@@ -1,7 +1,7 @@
 import React from 'react'
 export const DashboardProgress = (props) => {
   let prefix = 'hi-progress'
-  const {percent: percentNum, text, status, radius: radiusThis, withOutText = false} = props
+  const { percent: percentNum, text, status, radius: radiusThis, withOutText = false } = props
   const percent = percentNum > 0 ? percentNum : 0
   const radius = radiusThis > 0 ? radiusThis : 40
   const totalRadiusWidth = radius + 10
@@ -24,7 +24,7 @@ export const DashboardProgress = (props) => {
   return (
     <div
       className={`${prefix}__svg ${prefix}__svg_dashboard`}
-      style={{width: radius * 2, height: radius * 2}}>
+      style={{ width: radius * 2, height: radius * 2 }}>
       <svg
         viewBox={`0 0 ${totalRadiusWidth * 2} ${totalRadiusWidth * 2}`}>
         <path
