+++ components/progress/progress-legacy/style/index.js
